fix(lint): fix remaining implicit-widening in create and merge
Missed two spots in the previous lint fix: 4u * 1024u in create.c's
KB path, and (size_t)n * 1024u in merge.c's KB path.  Cast all
operands to size_t before multiplication.

+++ cmd/image/create.c
@@ -150,7 +150,7 @@ static size_t flash_size_bytes(const char *s)
 	if (errno != 0 || end == s || strcmp(end, "MB") != 0) {
 		/* Also accept sub-MB for bootloader pad-to-size (4KB). */
 		if (strcmp(s, "4KB") == 0)
-			return 4u * 1024u;
+			return (size_t)4u * (size_t)1024u;
 		die("unrecognised flash size '%s' (expected e.g. 2MB or 4KB)",
 		    s);
 	}
